+++ cmp/grid/columns/Column.js
@@ -185,7 +185,6 @@ export class Column {
                 hide: this.hidden,
                 minWidth: this.minWidth,
                 maxWidth: this.maxWidth,
-                rowHeight: this.rowHeight,
                 suppressResize: !this.resizable,
                 suppressMovable: !this.movable,
                 suppressSorting: !this.sortable,
@@ -282,19 +281,19 @@ export class Column {
 /**
  * @callback Column~rendererFn - normalized renderer function for a grid cell.
  * @param {*} value - cell data value (column + row).
- * @param {CellRendererMetadata} metadata - additional data about the column, row and GridModel.
+ * @param {CellRendererContext} context - additional data about the column, row and GridModel.
  * @return {string} - the formatted value for display.
  */
 
 /**
  * @callback Column~elementRendererFn - renderer function for a grid cell which returns a React component
  * @param {*} value - cell data value (column + row).
- * @param {CellRendererMetadata} metadata - additional data about the column, row and GridModel.
+ * @param {CellRendererContext} context - additional data about the column, row and GridModel.
  * @return {Element} - the React element to render.
  */
 
 /**
- * @typedef {Object} CellRendererMetadata
+ * @typedef {Object} CellRendererContext
  * @property {Record} record - row-level data Record.
  * @property {Column} column - column for the cell being rendered.
  * @property {GridModel} gridModel - gridModel for the grid.