+++ CHANGELOG.md
@@ -12,7 +12,7 @@ This is a desktop-only component.  Apps will have to adjust these imports.
 * Added `multiFieldRenderer`, an elementRendererFn that renders a collection of additional sub fields in a row beneath the main column field.
 Columns must define a `multiFieldRendererCfg` to use this renderer, which specifies which `subFields` to render.
 See MultiFieldRenderer for more information.
-* `Column.renderer` and `Column.elementRenderer` now receive `GridModel` as part of the `CellRendererMetadata`.
+* `Column.renderer` and `Column.elementRenderer` now receive `GridModel` as part of the `CellRendererContext`.
 * Column supports an optional `rowHeight` property, which specifies the cell height required (in pixels) by this columns renderer.
 Grids use this to determine an appropriate minimum row height when the column is visible.